admin_bot/requirements.txt
python-telegram-bot==21.6
boto3==1.35.0
python-dotenv==1.0.1
